Changes to Data
Data is now more realistic… for example, Pipe Material is now a drop
down with string values rather than an int value. Added error handling
for SchemaExceptions (Where the map tags are missing or wrong), and
InvalidInputException (Where the values are the wrong type). These are
thrown from the call() method in DataTask.